Enable bot to ignore messages from specified senders (#29)
* fix incorrect test

If there are two plugin functions listening to the same regex pattern, only the latest funciton will be loaded in plugin manager. Details can be found in bot.listen_to(). So there sould not be test for loading two plugin functions listening to the same regex pattern.

* test for 1 to 1 function-regexp mapping

append test to ensure the latest loaded function replaces previous loaded function which listening to the same regexp statement.

* unit test for ignoring senders specified in settings.IGNORE_USERS

* enable dispatcher to ignore messages from specified senders

* move get_sender as static method